rag/requirements.txt
langchain==0.3.27
langchain-google-genai==2.0.10
langchain-community==0.3.29
langchain-text-splitters==0.3.11
faiss-cpu==1.12.0
google-generativeai==0.8.5
google-genai==1.36.0
python-dotenv==1.1.1
sentence-transformers